rename TransporterSearch to HandlerSearch, move and change URL to /trak/handler/search to be reused for TSDF (& in future generator)
