+++ server/apps/trak/views/__init__.py
@@ -1,4 +1,3 @@
-from .handler import HandlerView
+from .handler import HandlerSearch, HandlerView
 from .manifest import ManifestView
 from .sites import SiteApi, SiteList, SiteManifest
-from .transporter import TransporterSearch
