+++ server/apps/trak/views/handler.py
@@ -1,4 +1,7 @@
+from http import HTTPStatus
+
 from django.core.exceptions import ObjectDoesNotExist
+from drf_spectacular.utils import extend_schema, inline_serializer
 from rest_framework import permissions, status
 from rest_framework.exceptions import APIException
 from rest_framework.generics import GenericAPIView
@@ -28,3 +31,30 @@ class HandlerView(GenericAPIView):
         except ObjectDoesNotExist:
             return self.response(status=status.HTTP_404_NOT_FOUND,
                                  data={'Error': f'{epa_id} not found'})
+
+
+class HandlerSearch(GenericAPIView):
+    queryset = Handler.objects.all()
+
+    @extend_schema(
+        description='Search for Transporters saved to the Haztrak database',
+        # methods=['POST'],
+        request=inline_serializer(name='test',
+                                  fields={'epaId': CharField(), 'name': CharField()}),
+    )
+    def post(self, request):
+        try:
+            epa_id = self.request.data['epaId']
+            name = self.request.data['name']
+            if len(epa_id) < 3 and len(name) < 3:
+                return Response(status=HTTPStatus.UNPROCESSABLE_ENTITY)
+            transporters_queryset = Handler.objects.filter(
+                site_type='Transporter').filter(
+                epa_id__icontains=epa_id).filter(name__icontains=name)
+            data = list(transporters_queryset)
+            response = []
+            for i in data:
+                response.append(HandlerSerializer(i).data)
+            return Response(status=HTTPStatus.OK, data=response)
+        except KeyError:
+            return Response(status=HTTPStatus.BAD_REQUEST)
